update csv timestamp only when required in bundle
This PR includes the following changes

add a script not to update the timestamp in the csv if there are no bundle changes
add ci action to verify bundle changes

With operator-sdk 1.26+ every time we generate a new bundle, the createdAt field
in the csv is updated.
This means we have to create a manual backport even when there are no bundle changes hence,
resetting the createdAt field if there are no changes in the bundle.

This PR also introduces a ci action to check whether there are any uncommitted bundle changes.

Reference from https://github.com/red-hat-storage/ocs-client-operator/pull/199

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>